leak norm ok on export, unset
